index.html
The Expense Tracker in JavaScript keeps a track record of expenditures. This is a simple lightweight program for personal expense manager in the web version. To run this project, first, you need to have a browser and a code editor of any kind. Then you have to run the index file from the editor or directly from the folder. When the screen loads, you can see the homepage. From there you can manage your expense activities.